+++ cmd/editMultiple.go
@@ -125,6 +125,16 @@ Start and end fields can't be mass-edited.`,
 			}
 		}
 
+		var dc *descriptionCompleter
+		if viper.GetBool(DESCR_AUTOCOMP) {
+			dc = newDescriptionCompleter(
+				c,
+				tei.WorkspaceID,
+				tei.UserID,
+				viper.GetInt(DESCR_AUTOCOMP_DAYS),
+			)
+		}
+
 		return manageEntry(
 			c,
 			tei,
@@ -151,6 +161,7 @@ Start and end fields can't be mass-edited.`,
 			},
 			shouldValidate,
 			false,
+			dc,
 		)
 	}),
 }